Adds S3 username translation support
This adds the ability to translate usernames following a json map in
a specified S3 bucket/file.

The new environment variables/options:
PLUGIN_S3_BUCKET and PLUGIN_S3_USERS_KEY can be set to an S3 bucket
and file key respectively. This will cause the plugin to load that
file and use its json string -> string object map to translate the
build.author and recipient fields into computed.authorSlack and
computed.recipientSlack.

These new fields can be used in templates normally. The
recipient option (to send a slack message to a specific user
vs a channel) with automatically translate the name when the
new S3 options are set.

The format of the JSON file:
{
   "gitUserName": "slackUserName",
}